Refactored PyAPI, again; removed 'tests' from PyAPI dir
