+++ tests/kv/test_save_load.py
@@ -15,17 +15,17 @@ class SaveLoad(NDBTest):
   async def test_save_load(self):
     input = {'a':0, 'b':'str', 'c':True, 'd':1.5}
 
-    await self.client.set(input)
+    await self.client.kv_set(input)
 
     datasetName = 'kv_'+ str(random.randint(1000,999999))
     
-    await self.client.save(datasetName)
+    await self.client.kv_save(datasetName)
 
-    clearedCount = await self.client.clear()
+    clearedCount = await self.client.kv_clear()
     self.assertEqual(clearedCount, len(input))
 
     # load
-    loadedCount = await self.client.load(datasetName)
+    loadedCount = await self.client.kv_load(datasetName)
     self.assertEqual(loadedCount, len(input))
