+++ tests/sessions/test_set_get.py
@@ -5,12 +5,12 @@ from base import NDBSessionTest
 class SetGet(NDBSessionTest):
 
   async def test_set(self):
-    session = await self.client.create_session()
+    session = await self.client.sh_create_session()
     self.assertTrue(session.isValid)
     
-    await self.client.set({'a':10, 'b':'x'}, session.tkn)
+    await self.client.sh_set({'a':10, 'b':'x'}, session.tkn)
     
-    info = await self.client.session_info(session.tkn)
+    info = await self.client.sh_info(session.tkn)
     self.assertTrue(session.isValid)
     self.assertEqual(info['tkn'], session.tkn)
     self.assertEqual(info['keyCnt'], 2)
@@ -24,15 +24,15 @@ class SetGet(NDBSessionTest):
     tokens = []
 
     for _ in range(0, nSessions):
-      session = await self.client.create_session()
+      session = await self.client.sh_create_session()
       self.assertTrue(session.isValid)
       
-      await self.client.set({'a':10, 'b':'x'}, session.tkn)
+      await self.client.sh_set({'a':10, 'b':'x'}, session.tkn)
 
       tokens.append(session.tkn)
 
 
-    info = await self.client.session_info_all()    
+    info = await self.client.sh_info_all()    
     self.assertEqual(info['totalSessions'], nSessions)
     self.assertEqual(info['totalKeys'], nSessions*2)  # we set 2 keys per session
