+++ tests/sv/test_server_info.py
@@ -5,7 +5,7 @@ from base import NDBTest
 class Info(NDBTest):
 
   async def test_info(self):
-    info = await self.client.server_info()
+    info = await self.client.sv_info()
     self.assertTrue('serverVersion' in info)
     self.assertEqual(info['persistEnabled'], True)
